Extracted KeyValueOption class
